backend/requirements.txt
aiohappyeyeballs         2.4.4
aiohttp                  3.11.11
aiosignal                1.3.2
alembic                  1.14.0
annotated-types          0.7.0
anyio                    4.7.0
attrs                    24.3.0
bcrypt                   4.0.1
certifi                  2024.12.14
charset-normalizer       3.4.1
click                    8.1.8
colorama                 0.4.6
dataclasses-json         0.6.7
dnspython                2.7.0
email_validator          2.2.0
fastapi                  0.115.6
fastapi-cli              0.0.7
filelock                 3.16.1
frozenlist               1.5.0
fsspec                   2024.12.0
greenlet                 3.1.1
h11                      0.14.0
httpcore                 1.0.7
httptools                0.6.4
httpx                    0.28.1
httpx-sse                0.4.0
huggingface-hub          0.27.1
idna                     3.10
importmonkey             2.1.1
iniconfig                2.0.0
itsdangerous             2.2.0
Jinja2                   3.1.5
joblib                   1.4.2
jose                     1.0.0
jsonpatch                1.33
jsonpointer              3.0.0
langchain-community      0.3.14
langchain-core           0.3.29
langchain-huggingface    0.1.2
langchain-text-splitters 0.3.5
langsmith                0.2.10
Mako                     1.3.8
markdown-it-py           3.0.0
MarkupSafe               3.0.2
marshmallow              3.25.1
mdurl                    0.1.2
mpmath                   1.3.0
multidict                6.1.0
my_project               0.1.0      D:\project-managemen-web-app\backend
mypy-extensions          1.0.0
networkx                 3.4.2
numpy                    1.26.4
orjson                   3.10.12
packaging                24.2
passlib                  1.7.4
pillow                   11.1.0
pip                      24.0
pluggy                   1.5.0
propcache                0.2.1
psycopg2                 2.9.10
psycopg2-binary          2.9.10
pydantic                 2.10.4
pydantic_core            2.27.2
pydantic-extra-types     2.10.1
pydantic-settings        2.7.0
Pygments                 2.18.0
PyJWT                    2.10.1
pytest                   8.3.4
python-dotenv            1.0.1
python-multipart         0.0.20
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
requests-toolbelt        1.0.0
rich                     13.9.4
rich-toolkit             0.12.0
safetensors              0.5.2
scikit-learn             1.6.1
scipy                    1.15.1
sentence-transformers    3.3.1
setuptools               70.0.0
shellingham              1.5.4
sniffio                  1.3.1
SQLAlchemy               2.0.36
sqlmodel                 0.0.22
starlette                0.41.3
sympy                    1.13.1
tenacity                 9.0.0
threadpoolctl            3.5.0
tokenizers               0.21.0
torch                    2.5.1
tqdm                     4.67.1
transformers             4.48.0
typer                    0.15.1
typing_extensions        4.12.2
typing-inspect           0.9.0
ujson                    5.10.0
urllib3                  2.3.0
uvicorn                  0.34.0
watchfiles               1.0.3
websockets               14.1
yarl                     1.18.3s